requirements.txt
matplotlib
numpy
python-dotenv